First stage refactor CredentitalManagers
