+++ src/Oauth2CredentialManagers/ArrayStore.php
@@ -7,41 +7,12 @@ use League\OAuth2\Client\Token\AccessTokenInterface;
 use Webfox\Xero\Oauth2Provider;
 use Webfox\Xero\OauthCredentialManager;
 
-class ArrayStore implements OauthCredentialManager
+class ArrayStore extends BaseCredentialManager implements OauthCredentialManager
 {
     public ?array $dataStorage = null;
 
     public function __construct(protected Store $session, protected Oauth2Provider $oauthProvider)
     {
-    }
-
-    public function getAccessToken(): string
-    {
-        return $this->data('token');
-    }
-
-    public function getRefreshToken(): string
-    {
-        return $this->data('refresh_token');
-    }
-
-    public function getTenants(): ?array
-    {
-        return $this->data('tenants');
-    }
-
-    public function getTenantId(int $tenant = 0): string
-    {
-        if (! isset($this->data('tenants')[$tenant])) {
-            throw new \Exception('No such tenant exists');
-        }
-
-        return $this->data('tenants')[$tenant]['Id'];
-    }
-
-    public function getExpires(): int
-    {
-        return $this->data('expires');
     }
 
     public function getState(): string
@@ -57,29 +28,9 @@ class ArrayStore implements OauthCredentialManager
         return $redirectUrl;
     }
 
-    public function getData(): array
-    {
-        return $this->data();
-    }
-
     public function exists(): bool
     {
         return $this->dataStorage !== null;
-    }
-
-    public function isExpired(): bool
-    {
-        return time() >= $this->data('expires');
-    }
-
-    public function refresh(): void
-    {
-        $newAccessToken = $this->oauthProvider->getAccessToken('refresh_token', [
-            'grant_type' => 'refresh_token',
-            'refresh_token' => $this->getRefreshToken(),
-        ]);
-
-        $this->store($newAccessToken);
     }
 
     public function store(AccessTokenInterface $token, array $tenants = null): void
@@ -91,26 +42,6 @@ class ArrayStore implements OauthCredentialManager
             'expires' => $token->getExpires(),
             'tenants' => $tenants ?? $this->getTenants(),
         ];
-    }
-
-    public function getUser(): ?array
-    {
-        try {
-            $jwt = new \XeroAPI\XeroPHP\JWTClaims();
-            $jwt->setTokenId($this->data('id_token'));
-            $decodedToken = $jwt->decode();
-
-            return [
-                'given_name' => $decodedToken->getGivenName(),
-                'family_name' => $decodedToken->getFamilyName(),
-                'email' => $decodedToken->getEmail(),
-                'user_id' => $decodedToken->getXeroUserId(),
-                'username' => $decodedToken->getPreferredUsername(),
-                'session_id' => $decodedToken->getGlobalSessionId(),
-            ];
-        } catch (\Throwable $e) {
-            return null;
-        }
     }
 
     protected function data(string $key = null)
