+++ src/Oauth2CredentialManagers/CacheStore.php
@@ -8,41 +8,12 @@ use League\OAuth2\Client\Token\AccessTokenInterface;
 use Webfox\Xero\Oauth2Provider;
 use Webfox\Xero\OauthCredentialManager;
 
-class CacheStore implements OauthCredentialManager
+class CacheStore extends BaseCredentialManager implements OauthCredentialManager
 {
     protected string $cacheKey = 'xero_oauth';
 
     public function __construct(protected Repository $cache, protected Store $session, protected Oauth2Provider $oauthProvider)
     {
-    }
-
-    public function getAccessToken(): string
-    {
-        return $this->data('token');
-    }
-
-    public function getRefreshToken(): string
-    {
-        return $this->data('refresh_token');
-    }
-
-    public function getTenants(): ?array
-    {
-        return $this->data('tenants');
-    }
-
-    public function getTenantId(int $tenant = 0): string
-    {
-        if (! isset($this->data('tenants')[$tenant])) {
-            throw new \Exception('No such tenant exists');
-        }
-
-        return $this->data('tenants')[$tenant]['Id'];
-    }
-
-    public function getExpires(): int
-    {
-        return $this->data('expires');
     }
 
     public function getState(): string
@@ -58,29 +29,9 @@ class CacheStore implements OauthCredentialManager
         return $redirectUrl;
     }
 
-    public function getData(): array
-    {
-        return $this->data();
-    }
-
     public function exists(): bool
     {
         return $this->cache->has($this->cacheKey);
-    }
-
-    public function isExpired(): bool
-    {
-        return time() >= $this->data('expires');
-    }
-
-    public function refresh(): void
-    {
-        $newAccessToken = $this->oauthProvider->getAccessToken('refresh_token', [
-            'grant_type' => 'refresh_token',
-            'refresh_token' => $this->getRefreshToken(),
-        ]);
-
-        $this->store($newAccessToken);
     }
 
     public function store(AccessTokenInterface $token, array $tenants = null): void
@@ -92,26 +43,6 @@ class CacheStore implements OauthCredentialManager
             'expires' => $token->getExpires(),
             'tenants' => $tenants ?? $this->getTenants(),
         ]);
-    }
-
-    public function getUser(): ?array
-    {
-        try {
-            $jwt = new \XeroAPI\XeroPHP\JWTClaims();
-            $jwt->setTokenId($this->data('id_token'));
-            $decodedToken = $jwt->decode();
-
-            return [
-                'given_name' => $decodedToken->getGivenName(),
-                'family_name' => $decodedToken->getFamilyName(),
-                'email' => $decodedToken->getEmail(),
-                'user_id' => $decodedToken->getXeroUserId(),
-                'username' => $decodedToken->getPreferredUsername(),
-                'session_id' => $decodedToken->getGlobalSessionId(),
-            ];
-        } catch (\Throwable $e) {
-            return null;
-        }
     }
 
     protected function data(string $key = null)
