+++ src/Oauth2CredentialManagers/FileStore.php
@@ -8,7 +8,7 @@ use League\OAuth2\Client\Token\AccessTokenInterface;
 use Webfox\Xero\Oauth2Provider;
 use Webfox\Xero\OauthCredentialManager;
 
-class FileStore implements OauthCredentialManager
+class FileStore extends BaseCredentialManager implements OauthCredentialManager
 {
     /** @var FilesystemManager */
     protected $disk;
@@ -20,35 +20,6 @@ class FileStore implements OauthCredentialManager
     {
         $this->disk = $files->disk(config('xero.credential_disk', config('filesystems.default')));
         $this->filePath = 'xero.json';
-    }
-
-    public function getAccessToken(): string
-    {
-        return $this->data('token');
-    }
-
-    public function getRefreshToken(): string
-    {
-        return $this->data('refresh_token');
-    }
-
-    public function getTenants(): ?array
-    {
-        return $this->data('tenants');
-    }
-
-    public function getTenantId(int $tenant = 0): string
-    {
-        if (! isset($this->data('tenants')[$tenant])) {
-            throw new \Exception('No such tenant exists');
-        }
-
-        return $this->data('tenants')[$tenant]['Id'];
-    }
-
-    public function getExpires(): int
-    {
-        return $this->data('expires');
     }
 
     public function getState(): string
@@ -64,29 +35,9 @@ class FileStore implements OauthCredentialManager
         return $redirectUrl;
     }
 
-    public function getData(): array
-    {
-        return $this->data();
-    }
-
     public function exists(): bool
     {
         return $this->disk->exists($this->filePath);
-    }
-
-    public function isExpired(): bool
-    {
-        return time() >= $this->data('expires');
-    }
-
-    public function refresh(): void
-    {
-        $newAccessToken = $this->oauthProvider->getAccessToken('refresh_token', [
-            'grant_type' => 'refresh_token',
-            'refresh_token' => $this->getRefreshToken(),
-        ]);
-
-        $this->store($newAccessToken);
     }
 
     public function store(AccessTokenInterface $token, array $tenants = null): void
@@ -101,26 +52,6 @@ class FileStore implements OauthCredentialManager
 
         if ($ret === false) {
             throw new \Exception("Failed to write to file: {$this->filePath}");
-        }
-    }
-
-    public function getUser(): ?array
-    {
-        try {
-            $jwt = new \XeroAPI\XeroPHP\JWTClaims();
-            $jwt->setTokenId($this->data('id_token'));
-            $decodedToken = $jwt->decode();
-
-            return [
-                'given_name' => $decodedToken->getGivenName(),
-                'family_name' => $decodedToken->getFamilyName(),
-                'email' => $decodedToken->getEmail(),
-                'user_id' => $decodedToken->getXeroUserId(),
-                'username' => $decodedToken->getPreferredUsername(),
-                'session_id' => $decodedToken->getGlobalSessionId(),
-            ];
-        } catch (\Throwable $e) {
-            return null;
         }
     }
